# api_go/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for SmartOrganizr.",
        "title": "SmartOrganizr API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/authors": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "List authors for the current user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page index (0-based)",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional name filter",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PagedAuthorRepresentationModelList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "Create an author",
                "parameters": [
                    {
                        "description": "New author payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthorCreateDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Author"
                        }
                    }
                }
            }
        },
        "/v1/authors/{authorId}": {
            "delete": {
                "tags": [
                    "authors"
                ],
                "summary": "Delete an author",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author ID",
                        "name": "authorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "Update an author",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author ID",
                        "name": "authorId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Author patch payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthorPatchDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Author"
                        }
                    }
                }
            }
        },
        "/v1/authors/{authorId}/notes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "List notes belonging to an author",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author ID",
                        "name": "authorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Note"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.Author": {
            "type": "object",
            "required": [
                "extraInformation",
                "id",
                "name"
            ],
            "properties": {
                "extraInformation": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.AuthorCreateDto": {
            "type": "object",
            "required": [
                "extraInformation",
                "name"
            ],
            "properties": {
                "extraInformation": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.AuthorPatchDto": {
            "type": "object",
            "required": [
                "extraInformation",
                "name"
            ],
            "properties": {
                "extraInformation": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.Page": {
            "type": "object",
            "required": [
                "number",
                "size",
                "totalElements",
                "totalPages"
            ],
            "properties": {
                "number": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "totalElements": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "dto.PagedAuthorRepresentationModelList": {
            "type": "object",
            "required": [
                "_embedded",
                "page"
            ],
            "properties": {
                "_embedded": {
                    "type": "object",
                    "required": [
                        "authorRepresentationModelList"
                    ],
                    "properties": {
                        "authorRepresentationModelList": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Author"
                            }
                        }
                    }
                },
                "page": {
                    "$ref": "#/definitions/dto.Page"
                }
            }
        },
...
        "models.Author": {
            "type": "object",
            "properties": {
                "extraInformation": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
        "models.Folder": {
            "type": "object",
            "properties": {
                "creationDate": {
                    "type": "string"
                },
                "creator": {
                    "$ref": "#/definitions/models.User"
                },
                "description": {
                    "type": "string"
                },
                "elements": {
                    "type": "array",
                    "items": {}
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "$ref": "#/definitions/models.Folder"
                }
            }
        },
        "models.Note": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.Author"
                },
                "creationDate": {
                    "type": "string"
                },
                "creator": {
                    "$ref": "#/definitions/models.User"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "numberOfPages": {
                    "type": "integer"
                },
                "parent": {
                    "$ref": "#/definitions/models.Folder"
                },
                "pdfAvailable": {
                    "type": "boolean"
                },
                "pdfContent": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "profilePicUrl": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "sideBarCollapsed": {
                    "type": "boolean"
                },
                "telephoneNumber": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}